Jarvis AI/JarvisLite/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
google-generativeai==0.3.1
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.14
httpx==0.25.1
